agent1/requirements.txt
yfinance>=0.2.32
supabase>=2.0.0
pandas>=2.1.0
python-dotenv>=1.0.0
APScheduler>=3.10.4
pytz>=2023.3
numpy>=1.24.0
